Bugfix: Map and Map of Map properties
Fixes simple properties where the value is a simple map but during
parsing the value is lost as it was assumed to be part of a function.

Fixes get_property/attribute of named map key. The map was returned
and not the specified key within the map. When a key is specified,
the key will be looked up instead of just returning the map.